pyproject.toml
[project]
name = "car-image-processor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.10.1",
    "bitsandbytes>=0.48.1",
    "clip",
    "huggingface-hub>=0.35.3",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "rembg>=2.0.67",
    "simple-aesthetics-predictor",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.0",
]

[tool.uv.sources]
clip = { git = "https://github.com/openai/CLIP.git" }
simple-aesthetics-predictor = { git = "https://github.com/shunk031/simple-aesthetics-predictor.git" }
